stop memory from blowing up on projects with compilation errors
In some case, a compilation failure in a package could cause
golangci-lint to repeatedly go through its dependencies, allocating
huge amounts of memory. Presumably at some point it would stop and
print the failure, but in my case it did not.

packages.Visit keeps track of the packages visited, so using that
instead prevents the memory allocation spiral.

Fixes #349