go.mod: Use `replace` to use upstream names when possible
For the most part, the version numbers in the `replace ()` stanza are
the same version numbers that were removed from the `require ()`
stanza.  There are 5 exceptions: 5 of the dependencies had been
patched to use the name `github.com/golangci/THING` instead of the
upstream name, this PR must undo that change.

In 2 of those cases, the change was in a stand-alone commit that was
the commit pointed to by `go.mod`; the lines for them have been
adjusted to point 1 commit prior:

	-github.com/golangci/dupl v0.0.0-20180902072040-3e9179ac440a
	+github.com/mibk/dupl => github.com/golangci/dupl v0.0.0-20180507061201-53b9af5a4536

	-github.com/golangci/gosec v0.0.0-20180901114220-66fb7fc33547
	+github.com/securego/gosec => github.com/golangci/gosec v0.0.0-20190211063915-c04360ff3e06

In the other 2 cases, the change had been made in the same way, but
was not the most recent commit; so I had to add a `git revert` commit:

	// https://github.com/golangci/errcheck/pull/3
	-github.com/golangci/errcheck v0.0.0-20181003203344-ef45e06d44b6
	+github.com/kisielk/errcheck => github.com/LukeShu/errcheck v0.0.0-20190519204902-14b08a384e90

	//https://github.com/golangci/go-tools/pull/1
	-github.com/golangci/go-tools v0.0.0-20180109140146-af6baa5dc196
	+honnef.co/go/tools => github.com/LukeShu/go-tools v0.0.0-20190519204515-e8d8fd50977c

In the final case, it was mixed in with another commit, so it wasn't
quite as simple as `git revert`:

	// https://github.com/golangci/gocyclo/pull/1
	-github.com/golangci/gocyclo v0.0.0-20180528134321-2becd97e67ee
	+github.com/fzipp/gocyclo => github.com/LukeShu/gocyclo v0.0.0-20190519210859-68da8eb3e8f7

Once those 3 PRs have been merged, "LukeShu" can be replaced with
"golangci" in `go.mod`.