satisfy the erris linter
Since erris was added in 67f22a4, we should set a good example and fixup
issues it has detected. Because errors.Is and errors.As were only added
in go1.13, go.mod has been updated to reflect this.

The only example is type switching over a returned error interface. This
is really gross with errors.As, so it has been left. It may be worth
following upstream.