Fix syntax highlighting for bash scripts
Fix syntax highlighting for bash scripts by changing language from
bash to sh in triple-backtick blocks: ```bash -> ```sh.

Some of these blocks had command output or prompt as part of the
listening which formatted in odd ways, e.g. colouring every occurrence
of the word "for" or "help" in purple.

With the sh language specification these wrong colours are gone, so use
it.