Use upstream version of prealloc
The forked version of prealloc at github.com/golangci/prealloc has not
been kept up to date with the latest changes from the upstream
repository, github.com/alexkohler/prealloc. This takes advantage of a
recent refactor in the upstream to allow importing the logic directly as
a package, which allows golangci to remove the overhead of maintaining a
fork while effectively bumping the version of prealloc used.

Fixes #991