Add "enable-new" command that enables any linters that are not explicitly disabled
This command updates the identified configuration file and adds new linters to
it, trying to preserve the original file as closely as possible

This offers an alternative to the deprecated "enable-all" setting.  It allows
projects to keep up-to-date with the latest linters without introducing new
linters when the golangci-lint version is upgraded.