Allow repeated importas aliases
The standalone importas takes arguments in the form "{package}:{alias}"
and stores them in a map in which package is the key. The intent is to
describe the one alias that is appropriate for a unique package path.

We cannot use the package path as a map key due to Viper's dot notation.
Instead use a list with package/alias pairs.