Allow errcheck exclude config without extra file
One of the challenges in standardizing on a golangci-lint configuration
is that if we want to configure `errcheck`, it requires two separate
files to configure: the golangci config file, plus a separate errcheck
excludes file. In particular, copying golangci-lint configs across
projects often fails, as a reference to an exclude file at a particular
file will not resolve.

While this config does not add any new functionality, it allows the
entirety of the errcheck configuration to be done inside the
golangci-lint config file itself.