gci: Parse the settings more similarly to the cli
The gci command line parses the -local flag using the gci.ParseLocalFlag
helper function, which is different than strings.Split (which
golangci-lint uses) in that it handles the empty string (which
golangci-lint separately handles) and that it handles empty entries in the
list (which golangci-lint does not handle).

For example, the gci command line parses "-local foo,,bar" as the list
["foo", "bar"], but golangci-lint parses the YAML
"local-prefixes: foo,,bar" as the list ["foo", "", "bar"].

This is a silly discrepancy to have; just use the gci.ParseLocalFlag
helper function!