runAnalyzersConfig: export analyzers
In order to use golangci-lint in non-trivial use cases, such as Bazel's
rules_go's 'nogo' static analysis framework, we would need to make the
analysis.Analyzer defined for various different golinters be reuseable
outside of golangci-lint.

Export the GetAnalyzers function of runAnalyzersConfig interface as well
as adjusting the 2 implementations of that interface: goanalysis.Linter
and goanalysis.MetaLinter

This should help enable other packages to be able to consumer
golangci-lint analyzers for different use cases.