Fix many linters ignoring Cgo files
This change fixes a bug caused that many linters ignored all files that are
using Cgo. It was introduced by PR #1065, which assumed that all files
referenced by //line directives are non-Go files, which is not true. In the case
of Cgo they point to the original Go files which are used by Cgo as templates to
generate other Go files.

The fix replaces all calls of Pass.Fset.PositionFor with a function
getFileNames() that first calls Pass.Fset.PositionFor with adjustment, and only
if it results in a non-Go file it calls Pass.Fset.PositionFor again without
adjustment.

Fixes: #2910

Signed-off-by: Sven Anderson <sven@anderson.de>