config: fix path matching when path prefix is used
When someone invokes golangci-lint in a sub-directory, they can set the path
prefix to make the output look like the invocation had been done in the
root. But this prefix was ignored when checking path patterns of exclude,
severity, and skip files/dirs entries, so those with matching for directories
only worked when golangci-lint was always invoked in the same directory.

The underlying problem is that the rules from the configuration get checked
before updating the path associated with the issues in the path fixer. To make
the prefix work, it now gets passed down into processors and added there to the
issue path before checking against the path pattern.

For exclude and severity rules, this could have been done through a separate
parameter, but bundling it in a new fsutils helper struct made the change a bit
smaller.