goheader: fix relative template path
go-header template-path setting is usually a relative filepath. As a
result, the linter only worked when golangci-lint was invoked from that
folder.

We detect if the template path is relative and prefix it with config
path.

go-header_bad.go test case was modified because old files are not
checked as seen here
https://github.com/denis-tingaikin/go-header/blob/v0.4.3/analyzer.go#L59-L63