Add funcguard linter
The linter can be configured to fail with a custom error message when
calls of specific functions or methods are found.
Functions or methods are identified by their full-path, including the
package and their type.

With the default configuration it detects usages of database/sql and
net/http functions without context parameter that have equivalent
functions that accept a context.