unexportedglobal: Linter to require '_' prefix on globals
This adds a unexportedglobal to golangci-lint.
[unexportedglobal][1] is a linter that requires that
unexported global variables and constants are prefixed with '_'.

  [1]: https://github.com/abhinav/unexportedglobal

```go
package foo

// Bad
var pool = sync.Pool{ /* ... */ }

// Good
var _pool = sync.Pool{ /* ... */ }
```

The idea is to eliminate the risk of conflict between
names of local variables and names of global variables.

The linter is inspired by the [Prefix Unexported Globals with `_`][2]
guidance in [Uber's Go Style Guide][3].

  [2]: https://github.com/uber-go/guide/blob/master/style.md#prefix-unexported-globals-with-_
  [3]: https://github.com/uber-go/guide/blob/master/style.md

The linter is not enabled by default.
I have also not added it to the 'style' preset
because not everyone may agree with this change.