Upgrade `paralleltest` linter from v1.0.7 to v1.0.8
Here, upgrade the `paralleltest` linter from v1.0.7 to v1.0.8. Full
release notes:

https://github.com/kunwardeep/paralleltest/releases/tag/v1.0.8

The main addition is a new `-ignoremissingsubtests` flag, which requires
that `t.Parallel()` be specified for top-level tests, but allows it to
be skipped for subtests. Here, we add a little golangci-lint infrastructure
to support it, along with some basic documentation.