build(docker): make binary versioning inside container consistent with goreleaser build
goreleaser currently strips the leading `v` off of the SemVer tag when passing to ldflags,
while the container build does not remove it.

Remove the leading `v`, if present, off of the version in container builds so that the
version returned by `golangci-lint --version` is consistent regardless of whether the
program is running standalone or in a container.