Add action to add checklist for new linter
This will add the checklist for new linters to the repository and
revision control. When the label `linter: new` is added the checklist
will automatically be added to the PR instead of having to copy-paste it
from an arbitrary source.

Since only collaborators and members can add labels it's not likely that
this will be abused, however to mitigate the risk even further the
action will reset the checklist comment if the tag is added and removed
multiple times.