feat: add `notag` linter
This add a new linter notag `github.com/guerinoni/notag`, which is able
to warn you about usage of tags in structs.
It can catches tags globally, by pkg name or by pkg path.

This is tested against this repo changing the config `.golangci.yml`
with this options
```
    notag:
      denied: "db"
      denied-pkg:
        - printers: xml
      denied-pkg-path:
        - github.com/golangci/golangci-lint/v2/pkg/config: validate
```

And the output looks ok.

Other tests are in the main repository.