pkg/goanalysis: expose underlying errors
If compilation or building of a Go package fails, export data will not
be available. Previously, we would mask the underlying failure and
display only a message that "no export data" was available. This has
led to lots of confusion among golangci-lint users (and users at my
company, which is how I discovered this problem).

golangci/golangci-lint#3996
golangci/golangci-lint#4630
golangci/golangci-lint#4912
golangci/golangci-lint#4964
golangci/golangci-lint#5037
golangci/golangci-lint#5051
golangci/golangci-lint#5184
golangci/golangci-lint#5428
golangci/golangci-lint#5437

In addition to numerous discussions. The only common thread between
the responses to these discussions, is that the "no export data" error
message is obscuring whatever the underlying problem is, and making it
more difficult to troubleshoot.

https://github.com/golangci/golangci-lint/discussions/2489
https://github.com/golangci/golangci-lint/discussions/2191
https://github.com/golangci/golangci-lint/discussions/2751
https://github.com/golangci/golangci-lint/discussions/3363
https://github.com/golangci/golangci-lint/discussions/4829
https://github.com/golangci/golangci-lint/discussions/4880

This fix simply walks the import tree and appends any found error
messages to the "no export data" error message. This will make it much
clearer for end users to discover the problem that is causing the
linter to fail.

Example output:

    ERRO Running error: can't run linter goanalysis_metalinter
    inspect: failed to load package errs: could not load export data: no export data for "github.com/segmentio/ctlstore/pkg/errs" because of error in imported package(s): error in package "github.com/segmentio/stats/v5": # github.com/segmentio/stats/v5
    vendor/github.com/segmentio/stats/v5/tag.go:38:35: in call to slices.IsSortedFunc, type func(a Tag, b Tag) int of tagCompare does not match inferred type func(a Tag, b Tag) bool for func(a E, b E) bool

Fixes #6056.