Support flag to clear outputs from config
This change set adds a new command line flag called
`--clear-config-outputs`. When setting this flag all configured outputs
from the config file will be cleared and only the ones from the command
line respected or if none, fallback to the default happens.

This is useful for applications like the golangci-lint LSP that invoke
golangci-lint and want to ensure that only the JSON output is enabled
without respecting the users config.